fix: credits_court variable type changed

+++ courtly/src/main/java/com/courtly/court/dto/CourtResponse.java
@@ -1,13 +1,11 @@
 package com.courtly.court.dto;
 
-import java.math.BigDecimal;
-
 public record CourtResponse(
 
     Long id,
     String name,
     String description,
-    BigDecimal creditsCost,
+    int creditsCost,
     boolean active,
     Long facilityId
 ) {
